Remove unneeded class
